Falta agregar reestriccion Se cambio tipos de datos de Stock de Long a Double por ello se modifico unos test. Se agrego los form de la tienda como su ABM

+++ src/main/java/ar/edu/utn/frba/dds/models/entities/recompensas/CalculadoraDePuntos.java
@@ -21,8 +21,8 @@ public class CalculadoraDePuntos {
   private static final double coeficienteHeladerasActivas =
       Double.parseDouble(coeficientesConfig.getProperty("puntaje.coeficiente.heladerasActivas"));
 
-  public static long calcular(Colaborador colaborador) {
-    return (long) (new DineroRepository().getTotal(colaborador) * coeficientePesosDonados +
+  public static Double calcular(Colaborador colaborador) {
+    return (new DineroRepository().getTotal(colaborador) * coeficientePesosDonados +
         new RedistribucionViandasRepository().getTotal(colaborador) * coeficienteViandasDistribuidas +
         (new DonacionViandasRepository().getTotal(colaborador)) * coeficienteViandasDonadas +
         (new EntregaTarjetasRepository().getTotal(colaborador)) * coeficienteTarjetasRepartidas +
